+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.4.1 [Unreleased] [Breacking Change]
+# 0.5.0
 - Changed a way how validatorjs parameters are passed to `reactiveMobXForm` creation function.
 - Changed a way how control names are build. from `person.0.firstName` to `person[0].firstname`
 - Introduced a `BaseControl` class to keep all logic shared between all types of Controls